requirements.txt
gunicorn
flask
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.23.1
pandas==1.4.3
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==2.2.1
zipp==3.8.1
click==8.1.3
colorama==0.4.5
textract
matplotlib
cloudinary 
flask-cors
python-dotenv
requests
pymongo==4.2.0